refac: abstract quiz class, assign agent to quiz class
